ChatBot1/backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
gunicorn
pdfminer.six
python-docx
docx2txt
sentence-transformers
torch